refactor: render tags
