+++ web/src/components/organisms/Project/Content/ContentList/index.tsx
@@ -1,4 +1,3 @@
-import Tag from "@reearth-cms/components/atoms/Tag";
 import ContentListMolecule from "@reearth-cms/components/molecules/Content/List";
 import CommentsPanel from "@reearth-cms/components/organisms/Common/CommentsPanel";
 import ViewsMenu from "@reearth-cms/components/organisms/Project/Content/ViewsMenu";
@@ -6,23 +5,6 @@ import ModelsMenu from "@reearth-cms/components/organisms/Project/ModelsMenu";
 import { useT } from "@reearth-cms/i18n";
 
 import useHooks from "./hooks";
-
-export const renderTags = (el: any, field: any) => {
-  type Tag = { id: string; name: string; color: string };
-  const tags: Tag[] | undefined = field.typeProperty?.tags;
-  const value: string = el.props.children as string;
-  const tagIds: Set<string> = new Set(value !== "-" ? value.split(", ") : []);
-  const filteredTags: Tag[] = tags?.filter((tag: Tag) => tagIds.has(tag.id)) || [];
-  return (
-    <>
-      {filteredTags.map(({ id, name, color }: Tag) => (
-        <Tag key={id} color={color.toLowerCase()}>
-          {name}
-        </Tag>
-      ))}
-    </>
-  );
-};
 
 const ContentList: React.FC = () => {
   const t = useT();
